requirements.txt
APScheduler==3.10.4
pymongo==4.6.3
Faker==24.8.0
python-dotenv==0.21.1